+++ netpyne/analysis.py
@@ -35,7 +35,7 @@ def plotData ():
             if kwargs == True: kwargs = {}
             elif kwargs == False: continue
             func = getattr(sim.analysis, funcName)  # get pointer to function
-            func(**kwargs)  # call function with user arguments
+            out = func(**kwargs)  # call function with user arguments
 
         # Print timings
         if sim.cfg.timing:
@@ -1338,7 +1338,7 @@ def plot2Dnet (include = ['allCells'], figSize = (12,12), showConns = True, save
 ######################################################################################################################################################
 ## Calculate normalized transfer entropy
 ######################################################################################################################################################
-def nTE(cells1 = [], cells2 = [], spks1 = None, spks2 = None, trange = [0,1000], binSize = 20, numShuffle = 30):
+def nTE(cells1 = [], cells2 = [], spks1 = None, spks2 = None, timeRange = None, binSize = 20, numShuffle = 30):
     ''' 
     Calculate normalized transfer entropy
         - cells1 (['all',|'allCells','allNetStims',|,120,|,'E1'|,('L2', 56)|,('L5',[4,5,6])]): Subset of cells from which to obtain spike train 1 (default: [])
@@ -1346,7 +1346,7 @@ def nTE(cells1 = [], cells2 = [], spks1 = None, spks2 = None, trange = [0,1000],
          subset of cells in network
         - spks1 (list): Spike train 1; list of spike times; if omitted then obtains spikes from cells1 (default: None)
         - spks2 (list): Spike train 2; list of spike times; if omitted then obtains spikes from cells2 (default: None)
-        - trange ([min, max]): Range of time to calculate nTE in ms (default: [0,1000])
+        - trange ([min, max]): Range of time to calculate nTE in ms (default: [0,cfg.duration])
         - binSize (int): Bin size used to convert spike times into histogram 
         - numShuffle (int): Number of times to shuffle spike train 1 to calculate TEshuffled; note: nTE = (TE - TEShuffled)/H(X2F|X2P)
 
@@ -1425,12 +1425,17 @@ def nTE(cells1 = [], cells2 = [], spks1 = None, spks2 = None, trange = [0,1000],
 
         spks2 = list(spkts)
 
+    # time range
+    if getattr(sim, 'cfg', None):
+        timeRange = [0,sim.cfg.duration]
+    else:
+        timeRange = [0, max(spks1+spks2)]
 
     inputVec = h.Vector()
     outputVec = h.Vector()
-    histo1 = histogram(spks1, bins = np.arange(trange[0], trange[1], binSize))
+    histo1 = histogram(spks1, bins = np.arange(timeRange[0], timeRange[1], binSize))
     histoCount1 = histo1[0] 
-    histo2 = histogram(spks2, bins = np.arange(trange[0], trange[1], binSize))
+    histo2 = histogram(spks2, bins = np.arange(timeRange[0], timeRange[1], binSize))
     histoCount2 = histo2[0] 
 
     inputVec.from_python(histoCount1)
@@ -1443,42 +1448,136 @@ def nTE(cells1 = [], cells2 = [], spks1 = None, spks2 = None, trange = [0,1000],
 ######################################################################################################################################################
 ## Calculate granger causality
 ######################################################################################################################################################
-def granger(spk1, spk2, binSize=5, trange=[0,2000]):
-    """
-    Typical usage is as follows:
-    from bsmart import pwcausalr
-    F,pp,cohe,Fx2y,Fy2x,Fxy=pwcausalr(x,ntrls,npts,p,fs,freq);
+def granger(cells1 = [], cells2 = [], spks1 = None, spks2 = None, label1 = 'spkTrain1', label2 = 'spkTrain2', timeRange = None, binSize=5, plotFig = True, 
+    saveData = None, saveFig = None, showFig = True):
+    ''' 
+    Calculate and optionally plot Granger Causality 
+        - cells1 (['all',|'allCells','allNetStims',|,120,|,'E1'|,('L2', 56)|,('L5',[4,5,6])]): Subset of cells from which to obtain spike train 1 (default: [])
+        - cells2 (['all',|'allCells','allNetStims',|,120,|,'E1'|,('L2', 56)|,('L5',[4,5,6])]): Subset of cells from which to obtain spike train 1 (default: [])
+         subset of cells in network
+        - spks1 (list): Spike train 1; list of spike times; if omitted then obtains spikes from cells1 (default: None)
+        - spks2 (list): Spike train 2; list of spike times; if omitted then obtains spikes from cells2 (default: None)
+        - label1 (string): Label for spike train 1 to use in plot
+        - label2 (string): Label for spike train 2 to use in plot
+        - timeRange ([min, max]): Range of time to calculate nTE in ms (default: [0,cfg.duration])
+        - binSize (int): Bin size used to convert spike times into histogram 
+        - plotFug (True|False): Whether to plot a figure showing Granger Causality Fx2y and Fy2x
+        - saveData (None|'fileName'): File name where to save the final data used to generate the figure (default: None)
+        - saveFig (None|'fileName'): File name where to save the figure;
+            if set to True uses filename from simConfig (default: None)(default: None)
+        - showFig (True|False): Whether to show the figure or not;
+            if set to True uses filename from simConfig (default: None)
 
-    Outputs:
-        F is the frequency vector for the remaining quantities
-    pp is the spectral power
-    cohe is the coherence
-    Fx2y is the causality of channel X to channel Y
-    Fy2x is the causality of channel Y to channel X
-    Fxy is the "instantaneous" causality (cohe-Fx2y-Fy2x I think)
-    Inputs:
-    x is the data for at least two channels, e.g. a 2x8000 array consisting of two LFP time series
-    ntrls is the number of trials (whatever that means -- just leave it at 1)
-    npts is the number of points in the data (in this example, 8000)
-    p is the order of the polynomial fit (e.g. 10 for a smooth fit, 20 for a less smooth fit)
-    fs is the sampling rate (e.g. 200 Hz)
-    freq is the maximum frequency to calculate (e.g. fs/2=100, which will return 0:100 Hz)
-    """
+        - Returns 
+            F: list of freqs
+            Fx2y: causality measure from x to y 
+            Fy2x: causality from y to x 
+            Fxy: instantaneous causality between x and y 
+            fig: Figure handle 
+    '''
     
-    from pylab import histogram, plot, show
     import numpy as np
-    from support.bsmart import pwcausalr
+    from netpyne.support.bsmart import pwcausalr
 
-    histo1 = histogram(spk1, bins = np.arange(trange[0], trange[1], binSize))
+    if not spks1:  # if doesnt contain a list of spk times, obtain from cells specified
+        cells, cellGids, netStimPops = getCellsInclude(cells1)
+        numNetStims = 0
+
+        # Select cells to include
+        if len(cellGids) > 0:
+            try:
+                spkts = [spkt for spkgid,spkt in zip(sim.allSimData['spkid'],sim.allSimData['spkt']) if spkgid in cellGids]
+            except:
+                spkts = []
+        else: 
+            spkts = []
+
+        # Add NetStim spikes
+        spkts = list(spkts)
+        numNetStims = 0
+        for netStimPop in netStimPops:
+            if 'stims' in sim.allSimData:
+                cellStims = [cellStim for cell,cellStim in sim.allSimData['stims'].iteritems() if netStimPop in cellStim]
+                if len(cellStims) > 0:
+                    spktsNew = [spkt for cellStim in cellStims for spkt in cellStim[netStimPop] ]
+                    spkts.extend(spktsNew)
+                    numNetStims += len(cellStims)
+
+        spks1 = list(spkts)
+
+    if not spks2:  # if doesnt contain a list of spk times, obtain from cells specified
+        cells, cellGids, netStimPops = getCellsInclude(cells2)
+        numNetStims = 0
+
+        # Select cells to include
+        if len(cellGids) > 0:
+            try:
+                spkts = [spkt for spkgid,spkt in zip(sim.allSimData['spkid'],sim.allSimData['spkt']) if spkgid in cellGids]
+            except:
+                spkts = []
+        else: 
+            spkts = []
+
+        # Add NetStim spikes
+        spkts = list(spkts)
+        numNetStims = 0
+        for netStimPop in netStimPops:
+            if 'stims' in sim.allSimData:
+                cellStims = [cellStim for cell,cellStim in sim.allSimData['stims'].iteritems() if netStimPop in cellStim]
+                if len(cellStims) > 0:
+                    spktsNew = [spkt for cellStim in cellStims for spkt in cellStim[netStimPop] ]
+                    spkts.extend(spktsNew)
+                    numNetStims += len(cellStims)
+
+        spks2 = list(spkts)
+
+
+    # time range
+    if timeRange is None:
+        if getattr(sim, 'cfg', None):
+            timeRange = [0,sim.cfg.duration]
+        else:
+            timeRange = [0, max(spks1+spks2)]
+
+    histo1 = histogram(spks1, bins = np.arange(timeRange[0], timeRange[1], binSize))
     histoCount1 = histo1[0] 
 
-    histo2 = histogram(spk2, bins = np.arange(trange[0], trange[1], binSize))
+    histo2 = histogram(spks2, bins = np.arange(timeRange[0], timeRange[1], binSize))
     histoCount2 = histo2[0] 
 
     fs = 1000/binSize
     F,pp,cohe,Fx2y,Fy2x,Fxy = pwcausalr(np.array([histoCount1, histoCount2]), 1, len(histoCount1), 10, fs, fs/2)
 
-    return F, Fx2y[0],Fy2x[0], Fxy[0]
+
+    # plot granger
+    fig = -1
+    if plotFig:
+        fig = figure()
+        plot(F, Fy2x[0], label = label2 + ' -> ' + label1)
+        plot(F, Fx2y[0], 'r', label = label1 + ' -> ' + label2)
+        xlabel('Frequency (Hz)')
+        ylabel('Granger Causality')
+        legend()
+        
+        # save figure data
+        if saveData:
+            figData = {'cells1': cells1, 'cells2': cells2, 'spks1': cells1, 'spks2': cells2, 'binSize': binSize, 'Fy2x': Fy2x[0], 'Fx2y': Fx2y[0], 
+            'saveData': saveData, 'saveFig': saveFig, 'showFig': showFig}
+        
+            _saveFigData(figData, saveData, '2Dnet')
+     
+        # save figure
+        if saveFig: 
+            if isinstance(saveFig, basestring):
+                filename = saveFig
+            else:
+                filename = sim.cfg.filename+'_'+'2Dnet.png'
+            savefig(filename)
+
+        # show fig 
+        if showFig: _showFigure()
+
+    return F, Fx2y[0],Fy2x[0], Fxy[0], fig